1. Добавил всю логику войд зон (класс, спрайты, анимацию, взрыв, урон и пр.) 2. Переделал замедление времени, теперь замеделение влияет действительно на всё (в том числе новоявленные войд зоны), кроме игрока

+++ Assets/Resources/Scripts/HealthPanel.cs
@@ -9,7 +9,7 @@ public class HealthPanel : MonoBehaviour
     public Image healthSlider;
     public List<Text> damageText;
 
-    public void HitFunction(float fillAmount, int damage)
+    public void HitFunction(float fillAmount, float damage)
     {
         if (fillAmount < 0) fillAmount = 0;
         healthSlider.fillAmount = fillAmount;
